Batched circulant transformation
- Enable to have different phases per batch. Still, only ndim=3 supported.
- Moved helper function to be able to reuse them.
