+++ tests/qttools/toeplitz/test_circulant.py
@@ -2,153 +2,19 @@
 
 import pytest
 
-from qttools import NDArray, xp
+from qttools import xp
 from qttools.toeplitz.circulant import (
     _get_dft_matrix,
     _get_idft_matrix,
+    _make_1D_block_circulant,
+    _make_2D_block_circulant,
+    _make_2D_block_phi_circulant,
     check_circulant,
     detransform_circulant,
     detransform_phi_circulant,
     transform_circulant,
     transform_phi_circulant,
 )
-
-
-def _make_1D_block_circulant(
-    a: NDArray,
-    sections: int,
-) -> NDArray:
-    """Helper function to transform a matrix into a block circulant matrix with
-    the given number of sections."""
-
-    if a.shape[-1] % sections != 0:
-        raise ValueError("The last dimension of a must be divisible by sections.")
-
-    if a.shape[-2] != a.shape[-1]:
-        raise ValueError(
-            "The second to last dimension of a must be equal to the last dimension of a."
-        )
-
-    block_size = a.shape[-1] // sections
-    # Take the first block-row (top n rows)
-    block_layer = a[..., :block_size, :]
-    blocks = xp.split(block_layer, sections, axis=-1)
-
-    matrix = xp.zeros_like(a)
-    for i in range(sections):
-        shifted_blocks = blocks[-i:] + blocks[:-i]
-        matrix[..., i * block_size : (i + 1) * block_size, :] = xp.concatenate(
-            shifted_blocks, axis=-1
-        )
-
-    return matrix
-
-
-def _make_2D_block_circulant(
-    a: NDArray,
-    sections_x: int,
-    sections_y: int,
-) -> NDArray:
-    """Helper function to transform a matrix into a block circulant matrix with
-    the given number of sections."""
-
-    if a.shape[-1] % sections_x != 0:
-        raise ValueError("The last dimension of a must be divisible by sections_x.")
-    if a.shape[-1] % sections_y != 0:
-        raise ValueError("The last dimension of a must be divisible by sections_y.")
-    if a.shape[-1] % (sections_x * sections_y) != 0:
-        raise ValueError(
-            "The last dimension of a must be divisible by the section product."
-        )
-
-    if a.shape[-2] != a.shape[-1]:
-        raise ValueError(
-            "The second to last dimension of a must be equal to the last dimension of a."
-        )
-
-    block_size_x = a.shape[-1] // sections_x
-
-    # make first circulant in the y direction
-    for i in range(0, a.shape[-1], block_size_x):
-        a[..., :block_size_x, i : i + block_size_x] = _make_1D_block_circulant(
-            a[..., :block_size_x, i : i + block_size_x],
-            sections=sections_y,
-        )
-
-    return _make_1D_block_circulant(a, sections=sections_x)
-
-
-def _make_1D_block_phi_circulant(
-    a: NDArray,
-    phase: complex,
-    sections: int,
-) -> NDArray:
-    """Helper function to transform a matrix into a block circulant matrix with
-    the given number of sections."""
-
-    if a.shape[-1] % sections != 0:
-        raise ValueError("The last dimension of a must be divisible by sections.")
-
-    if a.shape[-2] != a.shape[-1]:
-        raise ValueError(
-            "The second to last dimension of a must be equal to the last dimension of a."
-        )
-
-    block_size = a.shape[-1] // sections
-    # Take the first block-row (top n rows)
-    block_layer = a[..., :block_size, :]
-    blocks = xp.split(block_layer, sections, axis=-1)
-
-    matrix = xp.zeros_like(a)
-    for i in range(sections):
-        if i == 0:
-            phased_blocks = blocks[-i:]
-        else:
-            phased_blocks = [phase * block for block in blocks[-i:]]
-
-        shifted_blocks = phased_blocks + blocks[:-i]
-        matrix[..., i * block_size : (i + 1) * block_size, :] = xp.concatenate(
-            shifted_blocks, axis=-1
-        )
-
-    return matrix
-
-
-def _make_2D_block_phi_circulant(
-    a: NDArray,
-    phase_x: complex,
-    phase_y: complex,
-    sections_x: int,
-    sections_y: int,
-) -> NDArray:
-    """Helper function to transform a matrix into a block circulant matrix with
-    the given number of sections."""
-
-    if a.shape[-1] % sections_x != 0:
-        raise ValueError("The last dimension of a must be divisible by sections_x.")
-    if a.shape[-1] % sections_y != 0:
-        raise ValueError("The last dimension of a must be divisible by sections_y.")
-    if a.shape[-1] % (sections_x * sections_y) != 0:
-        raise ValueError(
-            "The last dimension of a must be divisible by the section product."
-        )
-
-    if a.shape[-2] != a.shape[-1]:
-        raise ValueError(
-            "The second to last dimension of a must be equal to the last dimension of a."
-        )
-
-    block_size_x = a.shape[-1] // sections_x
-
-    # make first circulant in the y direction
-    for i in range(0, a.shape[-1], block_size_x):
-        a[..., :block_size_x, i : i + block_size_x] = _make_1D_block_phi_circulant(
-            a[..., :block_size_x, i : i + block_size_x],
-            phase_y,
-            sections=sections_y,
-        )
-
-    return _make_1D_block_phi_circulant(a, phase_x, sections=sections_x)
 
 
 def test_dft_matrix(block_sections: int):
@@ -257,8 +123,12 @@ def test_transform_2D_phi_circulant(
             (batch_size, block_size, block_size)
         )
 
-    phase_x = xp.exp(2j * xp.pi / block_sections_x)
-    phase_y = xp.exp(2j * xp.pi / block_sections_y)
+    phase_x = xp.array(
+        [xp.exp(i * 2j * xp.pi / block_sections_x) for i in range(batch_size)]
+    )
+    phase_y = xp.array(
+        [xp.exp(i * 2j * xp.pi / block_sections_y) for i in range(batch_size)]
+    )
 
     a_circulant = _make_2D_block_phi_circulant(
         a,
